pyproject.toml
[project]
name = "ppt-image-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pytesseract>=0.3.13",
    "python-pptx>=1.0.2",
    "requests>=2.32.3",
]
